CLN refactor subprocess launch of python
